requirements.txt
asgiref==3.2.3
boto3==1.12.14
botocore==1.15.14
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.4
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-environ==0.4.5
django-heroku==0.3.1
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.9.5
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
